mv conversion func body to header
